requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
pillow>=10.1.0
numpy>=1.26.0
opencv-python>=4.8.1.78
pydantic>=2.5.0
python-dotenv>=1.0.0
ultralytics>=8.0.0
requests>=2.31.0